Many fixes.
Added the function to get single(sub-)domain certificates by using # in
front of the filename.
Fixed the problem that aliasses started with a number and gave an error
by adding 'a-' before the alias.
Changed the way certificates are generated by naming a certificate
different from the domain alias (yyyyMMddhhmmss instead of yyyyMMdd)